refactor: edit mobile ver
