+++ src/Data/dataMainEn.json
@@ -1,6 +1,6 @@
 {
   "header": {
-    "textNews": "NEWS",
+    "textNews": "EVENTS",
     "textMain": "HOME",
     "textBio": "BIOGRAPHY",
     "textWork": "WORKS <span class='accent-dot'>•</span> PROJECTS",
