+++ src/app/pages/Events/events.scss
@@ -1,31 +1,74 @@
 .news-section {
   display: grid;
   grid-template-columns: max-content max-content;
+  justify-content: center;
   row-gap: 80px;
   column-gap: 130px;
   margin-top: 210px;
 
 
-  @media (max-width: 1024px) {
+  @media (max-width: 1200px) {
     margin-top: 150px; // padding: 20px 30px 20px 30px;
     column-gap: 80px;
     row-gap: 60px;
+    margin-top: 90px;
   }
 
   @media (max-width: 700px) {
-
     display: flex;
     flex-direction: column;
+    gap: 20px;
   }
 
 
 }
 
+
+
+.pagination-container {
+  margin-top: 130px;
+  margin-bottom: calc(160px - 44px);
+
+
+  @media (max-width: 1200px) {
+    margin-top: 6px; // padding: 20px 30px 20px 30px;
+    margin-bottom: calc(60px - 44px);
+    row-gap: 60px;
+  }
+}
+
 .pagination {
+  position: relative;
   width: 100%;
   display: flex;
   justify-content: center;
   align-items: center;
+}
+
+.pagination::after {
+  content: '';
+  margin-left: 50px;
+  right: 0;
+  background-color: $dark;
+  width: 360px;
+  height: 1px;
+
+  @media (max-width: 1200px) {
+    display: none;
+  }
+}
+
+.pagination::before {
+  content: '';
+  margin-right: 50px;
+  left: 0;
+  background-color: $dark;
+  width: 360px;
+  height: 1px;
+
+  @media(max-width: 1200px) {
+    display: none;
+  }
 }
 
 .pagination__button {
@@ -48,6 +91,10 @@
   background-image: url('/00_news/00_TIME_LEFT_DARK.png');
   background-size: contain;
   background-repeat: no-repeat;
+
+  @media (max-width: 1200px) {
+    display: none;
+  }
 }
 
 .pagination__button--next {
@@ -57,4 +104,8 @@
   background-image: url('/00_news/00_TIME_RIGHT_DARK.png');
   background-size: contain;
   background-repeat: no-repeat;
+
+  @media (max-width: 1200px) {
+    display: none;
+  }
 }
