+++ src/style/header.scss
@@ -30,7 +30,7 @@
     display: flex;
     justify-content: space-between;
     margin: 35px 0 0px;
-    padding: 0 20px;
+    // padding: 0 20px;
     width: 100%;
 
     @media (max-width: 1280px) {
