+++ src/style/main.scss
@@ -28,5 +28,5 @@
 
 @import '/src/app/pages/FancyBox/fancybox.scss';
 
-@import '/src/app/pages/News/news.scss';
+@import '/src/app/pages/Events/events.scss';
 @import '/src/app/components/EventBlock/EventBlock.scss';
