+++ vite.config.js
@@ -28,7 +28,7 @@ export default defineConfig({
         contact: path.resolve('contact.html'),
         memorialObjects: path.resolve('memorial-objects.html'),
         rightToRest: path.resolve('right-to-rest.html'),
-        news: path.resolve('news.html'),
+        events: path.resolve('events.html'),
       },
       output: {
         // Опции выходных данных, позволяющие контролировать структуру имен файлов и др.
@@ -70,7 +70,7 @@ export default defineConfig({
       { from: /\/work-projects/, to: '/work-projects.html' },
       { from: /\/biography/, to: '/biography.html' },
       { from: /\/memorial-objects/, to: '/memorial-objects.html' },
-      { from: /\/news/, to: '/news.html' },
+      { from: /\/events/, to: '/events.html' },
     ],
   },
 });
